refactor: added params directory and added exports to services barrel export
