+++ src/app/services/index.ts
@@ -2,3 +2,4 @@ export * from './npm-registry';
 export * from './storage';
 export * from './date';
 export * from './error-handler';
+export * from './params';
