+++ src/app/services/params/index.ts
@@ -1,0 +1,1 @@
+export * from './params.service';
